no need for an int cast on line 7

+++ Program.cs
@@ -4,7 +4,7 @@ using OsuParsers.Enums;
 var osuReplay = ReplayDecoder.Decode(@"C:\Users\Plextora\Desktop\test.osr"); // osr to edit
 var mods = new List<Mods> { Mods.DoubleTime, Mods.Flashlight }; // list of mods to add for replay
 
-osuReplay.Mods = (int)Mods.None; // remove all mods from replay
+osuReplay.Mods = Mods.None; // remove all mods from replay
 foreach (var i in mods)
     osuReplay.Mods |= i; // add mods to replay
 osuReplay.Save(@"C:\Users\Plextora\Desktop\output.osr"); // edited osr output
